Make find vet page more responsive
