+++ src/components/App/App.css
@@ -9,6 +9,7 @@
 .emergency-instructions {
   width: 100%;
   height: 5vh;
+  min-height: 50px;
   display: flex;
   align-items: center;
   justify-content: center;
