+++ src/components/FindVet/FindVet.css
@@ -1,15 +1,19 @@
 .find-vet-page {
   width: 100vw;
-  height: 77vh;
+  height: 97vh;
   display: flex;
   align-items: center;
   justify-content: center;
+  flex-wrap: wrap;
+  overflow: scroll;
   background-color: #FFE7AD
 }
 
 .find-vet-form {
   width: 20%;
+  min-width: 400px;
   height: 60%;
+  min-height: 500px;
   display: flex;
   flex-direction: column;
   align-items: center;
