+++ src/components/FindVet/FindVet.js
@@ -17,12 +17,12 @@ const FindVet = () => {
   const selectedCity = useSelector(state => state.findVet.city)
   const cards = useSelector(state => state.findVet.cards).map(vet =>{
     return (
-      <Card className='vet-card' bg='warning'>
+      <Card className='vet-card' bg='warning' style={{'min-width': '300px', 'min-height': '350px'}}>
         <Card.Img variant='side' src={vet.attributes.image} className='vet-card-img mt-3'/>
         <Card.Body>
           <Card.Title className='card-title'>{vet.attributes.name}</Card.Title>
           <Card.Subtitle className='card-phone'>{vet.attributes.phone}</Card.Subtitle>
-          <Card.Text className='card.address'>{vet.attributes.address}</Card.Text>
+          <Card.Text className='card-address'>{vet.attributes.address}</Card.Text>
         </Card.Body>
       </Card>
     )
