+++ src/components/Header/Header.css
@@ -1,44 +1,22 @@
-header {
+
+.navbar {
+  width: 100vw;
+  height: 8vh;
+  min-height: 100px;
+  text-align: left;
+}
+
+.nav-container {
+  width: 20vw;
+  height: 8vh;
+  overflow: hidden;
+  display: flex;
+  align-items: flex-end;
+  
+}
+
+.nav {
   width: 100%;
-  height: 10vh;
-  display: flex;
-  align-items: flex-end;
-  justify-content: space-between;
-  background-color: white;
-}
-
-.header-logo {
-  width: 50vw;
-  height: 10vh;
-  display: flex;
-}
-
-.header-logo > h1 {
-  align-self: flex-end;
-  font-size: 3.5rem;
-}
-
-.header-img {
-  width: 5vw;
-  height: auto;
-  margin: 0% 0% 0% 2%
-}
-
-nav {
-  width: 25%;
-  height: 100%;
-  margin-right: 3%;
-  display: flex;
-  align-items: flex-end;
-  justify-content: space-between;
-}
-
-.nav-btn {
-  width: 6vw;
   height: 4vh;
-  margin-bottom: 8%;
-  border: solid black 2px;
-  border-radius: 6px;
-  background-color: white;
-  cursor: pointer;
+  margin-top: 3vh;
 }
