+++ src/components/LandingPage/LandingPage.css
@@ -1,6 +1,6 @@
 .landing-page {
   width: 100vw;
-  height: 77vh;
+  height: 87vh;
   background-color: #FFF3D6;
 }
