+++ src/components/LandingPage/LandingPage.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Modal from 'react-bootstrap/Modal'
+import LandingModal from '../LandingModal/LandingModal.js'
 
 import { updateSearchError, showModal } from '../../packages/redux/search-slice';
 import { createCards } from '../../packages/redux/cards-slice'
@@ -12,7 +12,6 @@ const LandingPage = () => {
   const cards = useSelector(state => state.cards.cards)
   const searchError = useSelector(state => state.search.error)
   const show = useSelector(state => state.search.show)
-  const modalItem = useSelector(state => state.search.item)
 
   const submitSearch = async e => {
     e.preventDefault()
@@ -32,13 +31,7 @@ const LandingPage = () => {
     }
   }
 
-  const handleClose = () => {
-    store.dispatch(showModal(false))
-  }
 
-  const capitalize = string => {
-    return string.charAt(0).toUpperCase() + string.slice(1)
-  }
 
   return (
     <main className='landing-page'>
@@ -59,18 +52,7 @@ const LandingPage = () => {
       <section className='item-cards-container'>
         {cards && cards}
       </section>
-      <Modal show={show} onHide={handleClose} centered>
-        <Modal.Header closeButton>
-          <Modal.Title>{capitalize(modalItem.attributes.name)}</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          <h5>Toxicity Level: {modalItem.attributes.toxicity} / 5</h5>
-          <h5>Description:</h5>
-          <p>{modalItem.attributes.description}</p>
-          <h5>Treatment:</h5>
-          <p>{modalItem.attributes.treatment}</p>
-        </Modal.Body>
-      </Modal>
+      <LandingModal show={show} />
     </main>
   )
 }
