--- src/components/Modal/Modal.js
@@ -1,16 +1,0 @@
-import React from 'react'
-import './Modal.css'
-
-const Modal = () => {
-
-  return (
-    <section className='modal-container'>
-      <article className='modal'>
-
-      </article>
-    </section>
-  )
-
-}
-
-export default Modal
