+++ src/components/SuggestionPage/SuggestionPage.css
@@ -1,6 +1,6 @@
 .suggestion-page {
   width: 100vw;
-  height: 77vh;
+  height: 79vh;
   background-color: #FFE7AD;
   display: flex;
   align-items: center;
